backend/requirements.txt
flask
supabase
python-dotenv
flask-cors
